General Updates - positioning
I've begun to add some more styling to the core parts of the website. Such as the sidebar nav menu, drop down menu, and the overall body of the page being centered.

I've also added border that are visible to the top bar and sidebar nav.

Fixed issue where the notification number/circle appeared distant when compared to the piece of text they're assigned to in the sidebar nav menu.